Ensure user has CAP_SYS_ADMIN in order to use mount
This only allows mount() to be used by the user with CAP_SYS_ADMIN in a
user namespace created by host root. i.e. it does not allow nesting.
Nesting could be supported, by checking the user namespace owns the
mount namespace, but this secures it against the obvious attack for now.